mpc: split out server config type and force ssl everywhere
