+++ src/mpc/coordinator/client.py
@@ -3,8 +3,9 @@
 from .crypto import \
     VerificationKey, Signature, export_digest, export_verification_key, \
     export_signature
-from requests import post, get
-from os.path import join
+from typing import Optional
+from requests import post, get, Response
+from os.path import join, exists
 import time
 
 CHUNK_SIZE = 4096
@@ -12,8 +13,10 @@ CHUNK_SIZE = 4096
 
 class Client(object):
 
-    def __init__(self, base_url: str):
+    def __init__(self, base_url: str, cert_path: Optional[str] = None):
+        assert not cert_path or exists(cert_path)
         self.base_url = base_url
+        self.cert_path = cert_path
 
     def get_challenge(self, challenge_file: str) -> None:
         """
@@ -24,8 +27,14 @@ class Client(object):
         # the next contributor, knowing his turn is next, to be waiting for
         # processing to complete.  Hence we loop with a message if the server
         # claims to be temporarily unavailable.
+        def _get_challenge() -> Response:
+            return get(
+                join(self.base_url, "challenge"),
+                stream=True,
+                verify=self.cert_path)
+
         while True:
-            with get(join(self.base_url, "challenge"), stream=True) as resp:
+            with _get_challenge() as resp:
                 if 503 == resp.status_code:
                     print("server is busy.  retrying ...")
                     time.sleep(5.0)
@@ -55,5 +64,6 @@ class Client(object):
             r = post(
                 join(self.base_url, "contribute"),
                 files={'response': upload_f},
-                headers=headers)
+                headers=headers,
+                verify=self.cert_path)
             r.raise_for_status()
